move lineLength, wordWrap, and warn to the Configuration class
