+++ lib/src/util/io.dart
@@ -18,29 +18,6 @@ const _newline = 0xA;
 /// The ASCII code for a carriage return character.
 const _carriageReturn = 0xD;
 
-/// The default line length for output when there isn't a terminal attached to
-/// stdout.
-final _defaultLineLength = int
-    .parse(Platform.environment['DEFAULT_TERMINAL_LINE_LENGTH'] ?? '200',
-        onError: (value) {
-  warn('Invalid value `$value` for DEFAULT_TERMINAL_LINE_LENGTH'
-      ', expected an int.');
-  return 200;
-});
-
-/// The maximum line length for output.
-final int lineLength = () {
-  try {
-    return stdout.terminalColumns;
-  } on UnsupportedError {
-    // This can throw an [UnsupportedError] if we're running in a JS context
-    // where `dart:io` is unavaiable.
-    return _defaultLineLength;
-  } on StdoutException {
-    return _defaultLineLength;
-  }
-}();
-
 /// The root directory of the Dart SDK.
 final String sdkDir = p.dirname(p.dirname(Platform.resolvedExecutable));
 
@@ -124,47 +101,6 @@ Stream<List<int>> sanitizeForWindows(Stream<List<int>> input) {
         .reversed
         .toList();
   });
-}
-
-/// Wraps [text] so that it fits within [lineLength].
-///
-/// This preserves existing newlines and doesn't consider terminal color escapes
-/// part of a word's length. It only splits words on spaces, not on other sorts
-/// of whitespace.
-String wordWrap(String text) {
-  return text.split("\n").map((originalLine) {
-    var buffer = new StringBuffer();
-    var lengthSoFar = 0;
-    for (var word in originalLine.split(" ")) {
-      var wordLength = withoutColors(word).length;
-      if (wordLength > lineLength) {
-        if (lengthSoFar != 0) buffer.writeln();
-        buffer.writeln(word);
-      } else if (lengthSoFar == 0) {
-        buffer.write(word);
-        lengthSoFar = wordLength;
-      } else if (lengthSoFar + 1 + wordLength > lineLength) {
-        buffer.writeln();
-        buffer.write(word);
-        lengthSoFar = wordLength;
-      } else {
-        buffer.write(" $word");
-        lengthSoFar += 1 + wordLength;
-      }
-    }
-    return buffer.toString();
-  }).join("\n");
-}
-
-/// Print a warning containing [message].
-///
-/// This automatically wraps lines if they get too long. If [color] is passed,
-/// it controls whether the warning header is color; otherwise, it defaults to
-/// [canUseSpecialChars].
-void warn(String message, {bool color}) {
-  if (color == null) color = canUseSpecialChars;
-  var header = color ? "\u001b[33mWarning:\u001b[0m" : "Warning:";
-  stderr.writeln(wordWrap("$header $message\n"));
 }
 
 /// Repeatedly finds a probably-unused port on localhost and passes it to
